requirements/prod.txt
jsonschema
sqlalchemy==0.9.8
psycopg2==2.5.4
alembic==0.7.0
anyjson==0.3.3
py-bcrypt==0.4
iptools==0.6.1
celery==3.1.17
pymemcache==1.3.3
redis==2.10.3
pyquery==1.2.9
raven[flask]

Flask==0.10.1
Flask-Login==0.2.11
Flask_OAuthlib==0.9.1
Flask-Migrate==1.2.0
Flask-Script==2.0.5
Flask_WTF==0.10.3
Flask-RESTful

version==0.1.1
pycrypto==2.6.1
Pillow==3.0.0
python-dateutil==2.3
psycogreen==1.0
jpush==3.2.1
git+https://github.com/zongxiao/slumber.git@master
dogpile.cache==0.5.4
qiniu==7.0.1
flower==0.7.3
qrcode==5.1
# python-weixin==0.1.5
ipython==2.4.0
xmltodict==0.9.2
lxml==3.6.0
elasticsearch==1.6.0
